Create README.md
Background: 
The client was an invoice financing company that helps businesses with cash flow by purchasing credit invoices at a discounted rate and taking the full amount when the invoice is paid by the customer. They approached us with a vast amount of data from the previous 3 years and wanted us to analyze it to derive insights that could help their business make better decisions.

Problem: 
The client didn't have a particular problem to solve, and we were afraid of getting stuck in analysis paralysis. We needed to have an in-depth discussion with the client to understand what pieces of analysis would be useful to them. Since the client wasn't familiar with the potential analyses, we had to do a lot of hand-holding to determine what needed to be analyzed.

Solution: 
After extensive discussion, we finalized the six points that we should analyze, which included monthly profit, overall efficiency, and distribution of clients by countries. Since the data didn't require a lot of pre-processing, we were able to turn it around in less than two days.

Outcome: 
The client was ecstatic with the results and realized the potential of the data. They gave us eight more problems to solve, including what-if analysis and time series machine learning models, which we completed without any issues. The client became one of our regular clients, and we continued to work together on multiple projects.